api/requirements.txt
ovmsclient
numpy
opencv-python
Pillow
Flask
flask-cors
